Add index.html, recipes/jollof.html and recipes/jollof.jpeg
index.html contain odin recipes which are recipes that I want to talk
about. In the index.html file, the first recipe that I write about which
is jollof rice have been linked and that is a separate director called
recipes. In the recipes directory, I have added an image of jollof rice
called jollof.jpeg.