Add content to yam-porridge.html and link it to styles.css
I added content to the yam-porridge.html and then added link the style
sheet to it. I had to put some styles and script in yam-porridge.html
to make some looks work well. I also had to change the link that linked
yam-porridge.html to index.html so that it work.