Edit images for consistency on the home page of website
The image cards on the home page are not looking equal, so I had to edit
the images for all of them to have an equal height and width. From there
I changed the name of the extensions in index.html so that they can
display on the home page, I will have to do the same for their various
pages too.