Touch fufu and yam porridge images a little bit and work on fufu.html
There were some white lines in the images after I first resized them,
so I went ahead to edit them again to get raid of those white lines.
Now I have change the extension of fufu image in fufu.html so that
the image of fufu will be displayed when that file is opened in
the browser.